backend/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# LLM + Agents
langchain==0.0.340

google-generativeai==0.3.0

# Embeddings
sentence-transformers==2.2.2

# Optional
requests==2.31.0
python-dotenv==1.0.0